+++ exe-unit/src/runtime/process.rs
@@ -14,8 +14,8 @@ use tokio::process::Command;
 use ya_agreement_utils::agreement::OfferTemplate;
 use ya_client_model::activity::{CommandOutput, ExeScriptCommand};
 use ya_manifest_utils::Feature;
-use ya_utils_process::{kill, ProcessTree, SystemError};
 use ya_runtime_api::server::{spawn, RunProcess, RuntimeControl, RuntimeService};
+use ya_utils_process::{kill, ProcessTree, SystemError};
 
 use crate::acl::Acl;
 use crate::error::Error;
