+++ utils/process/src/unix/usage.rs
@@ -1,4 +1,14 @@
+use std::mem;
+use std::time::Duration;
+
+use crate::SystemError;
+
 use super::Process;
+
+#[cfg(target_os = "linux")]
+use nix::unistd::sysconf;
+#[cfg(target_os = "linux")]
+use nix::unistd::SysconfVar::CLK_TCK;
 
 impl Process {
     pub fn usage(pid: i32) -> Result<Usage, SystemError> {
@@ -9,6 +19,15 @@ impl Process {
         let rss_gib = stat.rss as f64 / (1024. * 1024.);
 
         Ok(Usage { cpu_sec, rss_gib })
+    }
+
+    pub fn info(pid: i32) -> Result<Process, SystemError> {
+        let stat = StatStub::read(pid)?;
+        Ok(Process {
+            pid: stat.pid,
+            ppid: stat.ppid,
+            pgid: stat.pgid,
+        })
     }
 
     fn ticks_per_second() -> Result<i64, SystemError> {
@@ -116,5 +135,34 @@ impl StatStub {
         stub.rss = next(&mut it)?;
 
         Ok(stub)
+    }
+}
+
+#[cfg(test)]
+mod test {
+
+    #[cfg(target_os = "linux")]
+    #[test]
+    fn parse_stat() {
+        let stat = "10220 (proc-name) S 7666 7832 7832 0 -1 4194304 2266 0 4 0 44 2 0 0 \
+        20 0 6 0 1601 816193536 1793 18446744073709551615 94873375535104 94873375567061 \
+        140731153968032 0 0 0 0 4096 0 0 0 0 17 6 0 0 0 0 0 94873375582256 94873375584384 \
+        94873398587392 140731153974918 140731153974959 140731153974959 140731153977295 0";
+
+        let parsed = super::StatStub::parse_stat(stat).unwrap();
+        let expected = super::StatStub {
+            pid: 10220,
+            comm: "proc-name".to_string(),
+            state: 'S',
+            ppid: 7666,
+            pgid: 7832,
+            sid: 7832,
+            utime: 44,
+            stime: 2,
+            vsize: 816193536,
+            rss: 1793,
+        };
+
+        assert_eq!(parsed, expected);
     }
 }
